--- src/modules/api-tracking/api-traking.module.ts
@@ -1,19 +1,0 @@
-// Third Party Dependencies.
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-
-// Local Dependencies.
-import { ApiTracking } from './entities/api-traking.entity';
-import { ApiTrackingService } from './services/api-traking.service';
-import { ApiTrackingController } from './controllers/api-traking.controller';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([
-        ApiTracking,
-    ])
-  ],
-  providers: [ApiTrackingService],
-  controllers: [ApiTrackingController],
-})
-export class ApiTrakingModule {}
